# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TeamFlow API
  version: 0.1.0
  description: |
    Outline contract for the MVP. Cookie JWT auth (see docs/adr/0002-cookie-jwt-auth.md).
    Errors use RFC 7807 `application/problem+json`.
    Resource authorization is server-side; workspace role is not taken from the JWT.
  contact:
    name: TeamFlow
servers:
  - url: http://localhost:8080
    description: Local API (Compose or Spring Boot)
security:
  - cookieAuth: []
    csrfHeader: []
...
paths:
...
  /api/admin/users:
    get:
      tags: [admin]
      summary: List all active (non-deleted) users
      operationId: listUsers
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUserResponse"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
    post:
      tags: [admin]
      summary: Create a user (systemRole defaults to USER)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /api/admin/users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [admin]
      summary: Get a user
      operationId: getUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
    patch:
      tags: [admin]
      summary: Update name, email, system role and/or reset the password
      description: |
        Partial update; omitted fields are unchanged. No current password is required.
        Demoting the last system ADMIN is refused with 409.
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
    delete:
      tags: [admin]
      summary: Delete a user (soft delete + anonymize)
      description: |
        Same clean-up as self-service account deletion. Refused with 409 for the caller's own account
        (use DELETE /api/users/me) and while the user is the last ADMIN of a workspace that has other members.
      operationId: deleteUser
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
...
  responses:
    Problem:
      description: RFC 7807 error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    SystemRole:
      type: string
      description: Platform-wide role, independent of workspace roles
      enum: [ADMIN, USER]
...
    AdminUserResponse:
      type: object
      required: [id, email, displayName, systemRole, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        systemRole:
          $ref: "#/components/schemas/SystemRole"
        createdAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          minLength: 8
          maxLength: 128
        displayName:
          type: string
          maxLength: 80
        systemRole:
          $ref: "#/components/schemas/SystemRole"
    UpdateUserRequest:
      type: object
      description: Partial update; omitted fields are unchanged.
      properties:
        displayName:
          type: string
          maxLength: 80
        email:
          type: string
          format: email
          maxLength: 320
        systemRole:
          $ref: "#/components/schemas/SystemRole"
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: New password; sets it without needing the current one
...
    Problem:
      type: object
      required: [title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...